Move util functions into `/modules`
